Use FeatureWriter to enable triggered GPKG view creation
The FeatureWriter can be used to trigger creation of views in the
result GPKG. In this case a PythonCaller is used as the SQLExecutor
does not allow for INSERT statements.

Also split the InlineQuerier into a simpler InlineQuerier for only
SDT data and use a FeatureJoiner to reference SDT factors. This makes
for simpler attribute type normalization as the InlineQuerier is not
always updated correctly.

Finally, filter input/output attributes using a pair of AttributeKeepers.